requirements.txt
Django==5.2.6
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.6.0
pyotp==2.9.0
qrcode[pil]==8.0
Pillow==11.1.0
python-decouple==3.8
django-extensions==3.2.3
drf-spectacular==0.27.2
celery==5.3.6
redis==5.2.1
gunicorn==23.0.0
psycopg2-binary==2.9.10
whitenoise==6.8.2
django-crispy-forms==2.3
crispy-bootstrap5==2024.10
django-widget-tweaks==1.5.0
PyJWT==2.8.0